feat: move chart utils into shared file
